+++ CHANGELOG.md
@@ -4,6 +4,9 @@ CHANGELOG
 ## HEAD (Unreleased)
 _(none)_
 
+- feat(autoapi): add workspace scoped envvars to LocalWorkspace and Stack
+  [#5275](https://github.com/pulumi/pulumi/pull/5275)
+
 ## 2.9.2 (2020-08-31)
 
 - Alpha version of the Automation API for Go